move duplicate code to own file, differtiate between throwing and non throwing code
