+++ packages/data-mate/src/data-frame/search/interfaces.ts
@@ -1,7 +1,3 @@
 export interface MatchRowFn {
     (rowIndex: number): boolean;
 }
-
-export interface MatchValueFn {
-    (value: unknown): boolean;
-}
