+++ packages/data-mate/src/data-frame/search/wildcards-and-regex-utils.ts
@@ -2,7 +2,7 @@ import {
     wildCardToRegex, matchWildcard, match,
     isString, getTypeOf
 } from '@terascope/utils';
-import { MatchValueFn } from './interfaces.js';
+import { MatchValueFn } from '../../interfaces.js';
 
 export function regexp(regexStr: unknown): MatchValueFn {
     if (!isString(regexStr)) {
